Refactor FigureSupplier, changed int to double in subclass of the Figure.

+++ src/main/java/core/basesyntax/figures/Circle.java
@@ -1,9 +1,9 @@
 package core.basesyntax.figures;
 
 public class Circle extends Figure {
-    private final int radius;
+    private final double radius;
 
-    public Circle(String color, int radius) {
+    public Circle(String color, double radius) {
         super(color);
         this.radius = radius;
     }
